% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_tracks.R
\docType{data}
\name{ext_tracks}
\alias{ext_tracks}
\title{Tropical Storms dataset}
\format{
A \code{tibble} with 11824
observations and 29 columns
\describe{
#'   \item{storm_id}{chr, unique id specific for the storm}
\item{storm_name}{chr, name of the storm}
\item{month}{chr, month ormonths when the storm occurred}
\item{day}{chr, days the storm lasted}
\item{hour}{chr, hours the storm was recorded}
\item{year}{chr, year when storm occurred}
\item{longitude}{num, longitude of storm eye}
\item{latitude}{num, latitude of storm eye}
\item{max_wind}{num, maximum wind speed}
\item{min_pressure}{num, minimum pressure}
\item{rad_max_wind}{num, }
\item{eye_diameter}{num, }
\item{pressure_1}{num, }
\item{pressure_2}{num, }
\item{radius_34_ne}{num, }
\item{radius_34_se}{num, }
\item{radius_34_sw}{num, }
\item{radius_34_nw}{num, }
\item{radius_50_ne}{num, }
\item{radius_50_se}{num, }
\item{radius_50_sw}{num, }
\item{radius_50_nw}{num, }
\item{radius_64_ne}{num, }
\item{radius_64_se}{num, }
\item{radius_64_sw}{num, }
\item{radius_64_nw}{num, }
\item{storm_type}{chr, category of storm}
\item{distance_to_land}{num, shortest distance from storm eye to land area}
\item{final}{chr, when storm ended}
}
}
\source{
Extended Best Tract dataset,
\url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/}
}
\usage{
ext_tracks
}
\description{
Tropical Storms dataset
}
\examples{
\dontrun{
ext_tracks

data("ext_tracks")
}
}
\keyword{datasets}
