% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotdfss.R
\name{sotdfss}
\alias{sotdfss}
\title{sotdfss single observation time dataset for single storm}
\usage{
sotdfss(
  name = "KATRINA",
  yyyy = 2005,
  mm = "08",
  dd = "29",
  hh = "12",
  database = ext_tracks
)
}
\arguments{
\item{name}{character, storm name}

\item{yyyy}{character, storm year, four digits number as character}

\item{mm}{character, storm month, two digits number as character}

\item{dd}{character, storm day, two digits number as character}

\item{hh}{character, storm hour, two digits number as character}

\item{database}{tibble or data frame with storms information}
}
\value{
data frame with two variables: lon, lat
}
\description{
Create a single observation time dataset for a single storm when information
such as storm name, year, month, day, hour are provided
}
\examples{
\dontrun{
sotdfss("KATRINA", 2005, "08", "29", "12", ext_tracks)
}
}
