% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hurricane.R
\name{stat_hurricane}
\alias{stat_hurricane}
\title{Layer function for StatHurricane}
\usage{
stat_hurricane(
  mapping = NULL,
  data = NULL,
  geom = "hurricane",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale_radii = 1,
  np = 250,
  ...
)
}
\arguments{
\item{mapping, }{how variables in the data is mapped to visual properties}

\item{data, }{data used to create the graphical element}

\item{geom}{geom layer function to call}

\item{position}{where to plot each group of graphical element}

\item{na.rm}{parameter for na's}

\item{show.legend}{description to be displayed}

\item{inherit.aes}{logical aes parameter}

\item{scale_radii}{parameter used to scale wind radii}

\item{np}{number of points for each section: arc or segment}

\item{...}{other arguments to be passed}
}
\value{
graphical element
}
\description{
Compute layers to be added when mapping variables in data transformed by
StatHurricane to aesthetics to create plots
}
\examples{
\dontrun{
stat_hurricane()
}
}
