% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wborder.R
\name{wborder}
\alias{wborder}
\title{Coordinates of boundary points of storm}
\usage{
wborder(x0, y0, rne, rnw, rse, rsw, scaler = 1, np = 250)
}
\arguments{
\item{x0}{numeric, longitude of the eye of the storm}

\item{y0}{numeric, latitude of the eye of the storm}

\item{rne}{numeric, wind radius in North-East quadrant}

\item{rnw}{numeric, wind radius in North-West quadrant}

\item{rse}{numeric, wind radius in South-East quadrant}

\item{rsw}{numeric, wind radius in South-West quadrant}

\item{scaler}{numeric, to scale wind radii}

\item{np}{numeric, number of in each section: segment or arc}
}
\value{
data frame of two variables: lon, lat. Coordinates of wind boundary points
}
\description{
Compute boundary points for a single storm using center and radii in each quadrant
}
\examples{
\dontrun{
wborder(0, 0, 200, 100, 150, 75)
}
}
