% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wborders.R
\name{wborders}
\alias{wborders}
\title{Coordinates of boundaries points of storms wind radii}
\usage{
wborders(x0, y0, rne, rnw, rse, rsw, scaler = 1, np = 250)
}
\arguments{
\item{x0}{numeric vector containing longitudes of storm eyes}

\item{y0}{numeric vector containing latitudes of storm eyes}

\item{rne}{numeric vector containing storm radii in North-East quadrant}

\item{rnw}{numeric vector containing storm radii in North-West quadrant}

\item{rse}{numeric vector containing storm radii in South-East quadrant}

\item{rsw}{numeric vector containing storm radii in South-West quadrant}

\item{scaler}{numeric, a number between 0 and 1 to scale wind radii}

\item{np}{numeric, number of point in each section: arc or segment}
}
\value{
data frame of two variables lon and lat
}
\description{
Compute boundaries points for multiple singles storms using storm eyes and wind radii
}
\examples{
\dontrun{
wborders(c(0, 0), c(5, 7), c(400, 100), c(300, 200), c(200, 200), c(50, 50))
}
}
